LRQA-63828 Added in a function called dlhf for testers to utilize for testing hotfixes

+++ setup.sh
@@ -243,25 +243,10 @@ rstaging ()
 	notify-send "The bundle is ready for testing!"
 }
 
-dlhf ()
-{
-	#Check to see if the tester requeires a fixpack in their hotfix
-		read -p 'Does your hotfix require a fixpack? Please type y or n: ' fixpack
-
-	#Perform the respective functions that the tester needs
-		if [ $fixpack == "y" ]
-			then
-  				echo -e "\e[44mRunning fixpack download\e[0m"
-				  dlhffp
-		else
-  			echo  -e "\e[44mRunning non-fixpack download\e[0m"
-				dlhfnfp
-			fi
-}
-
-dlhffp()
+dlhf()
 {  
 # Clear out {username} and add your own username for this to work
+# Does not include putting the licence.xml and the mysql.jar
 # Clean out old copy
 	echo -e "\e[44mCleaning out old extracted binaries/folders\e[0m"
 	rm -rf $cwd/bundles/hotfixer
@@ -271,63 +256,16 @@ dlhffp()
 	mkdir -p bundles/hotfixer
 
 # Check download it
-   echo -e "\e[44mDownloading the master snapshot if its updated\e[0m"
-  read -p 'Please specify download link for portal: ' portaldl
-	wget -c -N --user=calum.ragan --ask-password $portaldl -P $cwd/bundles/hotfixer
-	7z x $cwd/bundles/hotfixer/*.7z -O$cwd/bundles/hotfixer
-	echo -e "\e[44mDone downloading or checking\e[0m"
-
-	cd $cwd/bundles/hotfixer/*/
-	echo -e 'Changed directory to liferay.home'
-	echo -e 'Current directory 2: ' $(pwd)
-  
-#  Download proper patching tool
-	echo -e "\e[44mDownloading the patching tool\e[0m"
-	rm -rf $cwd/bundles/hotfixer/*/patching-tool
-  read -p 'Please specify download link for patching tool: ' patchdl
-	wget -c -N --user=calum.ragan --ask-password $patchdl -P $cwd/bundles/hotfixer/*/
-	unzip $cwd/bundles/hotfixer/*/*.zip -d $cwd/bundles/hotfixer/*/
-	echo -e "\e[44mDone downloading or checking\e[0m"
-
-	#Change directory into patching tool folder
-	echo -e 'Current directory 3: ' $(pwd)
- 	cd $cwd/bundles/hotfixer/*/patching-tool
- 	echo -e 'changing into patching tool directory..' $(pwd)
-  
-  # Navigate into patching tool and run commands to install patch
-	echo -e "\e[44mRunning auto-discovery\e[0m"
-	 $cwd/bundles/hotfixer/*/patching-tool/patching-tool.sh auto-discovery 
-	echo -e "\e[44mDownloading the hotfix\e[0m"
-	read -p 'Please specify download link for the fixpack: ' fixpackdl
-	wget -c -N --user=calum.ragan --ask-password $fixpackdl -P $cwd/bundles/hotfixer/*/patching-tool/patches
-	read -p 'Please specify download link for the hotfix: ' hotfixdl
-	wget -c -N --user=calum.ragan --ask-password $hotfixdl -P $cwd/bundles/hotfixer/*/patching-tool/patches
-	$cwd/bundles/hotfixer/*/patching-tool/patching-tool.sh install 
-
-	# Deleting Data and Logs folders and cleaning DB
-	cd $cwd/bundles/hotfixer/*/
-	echo -e "\e[44mDeleting the logs and data folder\e[0m"
-	rm -r data
-	rm -r logs
-	echo -e "\e[44mRunning the cleandb command-1\e[0m"
-	cleandb
-}
-
-dlhfnfp(){  
-# Clear out {username} and add your own username for this to work
-# Clean out old copy
-	echo -e "\e[44mCleaning out old extracted binaries/folders\e[0m"
-	rm -rf $cwd/bundles/hotfixer
-
-# Create directories that will be used
-	echo -e "\e[44mCreating folders for bundle backup if they don't exist\e[0m"
-	mkdir -p bundles/hotfixer
-
-# Check download it
-   echo -e "\e[44mDownloading the master snapshot if its updated\e[0m"
-  read -p 'Please specify download link for portal: ' portaldl
-	wget -c -N --user=calum.ragan --ask-password $portaldl -P $cwd/bundles/hotfixer
-	7z x $cwd/bundles/hotfixer/*.7z -O$cwd/bundles/hotfixer
+   	echo -e "\e[44mDownloading the master snapshot if its updated\e[0m"
+   	read -p "Is the portal version a zip file? Please type y or n " portalver
+  	read -p 'Please specify download link for portal: ' portaldl
+	wget -c -N --user={username} --ask-password $portaldl -P $cwd/bundles/hotfixer
+	if [ $portalver == 'y' ]
+		then
+			unzip $cwd/bundles/hotfixer/*.zip -d $cwd/bundles/hotfixer/
+	else
+		7z x $cwd/bundles/hotfixer/*.7z -O$cwd/bundles/hotfixer
+	fi
 	echo -e "\e[44mDone downloading or checking\e[0m"
 
 	cd $cwd/bundles/hotfixer/*/
@@ -338,7 +276,7 @@ dlhfnfp(){
 	echo -e "\e[44mDownloading the patching tool\e[0m"
 	rm -rf $cwd/bundles/hotfixer/*/patching-tool
   	read -p 'Please specify download link for patching tool: ' patchdl
-	wget -c -N --user=calum.ragan --ask-password $patchdl -P $cwd/bundles/hotfixer/*/
+	wget -c -N --user={username} --ask-password $patchdl -P $cwd/bundles/hotfixer/*/
 	unzip $cwd/bundles/hotfixer/*/*.zip -d $cwd/bundles/hotfixer/*/
 	echo -e "\e[44mDone downloading or checking\e[0m"
 
@@ -350,19 +288,27 @@ dlhfnfp(){
   # Navigate into patching tool and run commands to install patch
 	echo -e "\e[44mRunning auto-discovery\e[0m"
 	 $cwd/bundles/hotfixer/*/patching-tool/patching-tool.sh auto-discovery 
+	 $cwd/bundles/hotfixer/*/patching-tool/patching-tool.sh revert
 	echo -e "\e[44mDownloading the hotfix\e[0m"
+	read -p "Does this hotfix require a fixpack? Please type y or n " fixpack
+	if [ $fixpack == 'y' ]
+		then
+			read -p 'Please specify download link for the fixpack: ' fixpackdl
+	wget -c -N --user={username} --ask-password $fixpackdl -P $cwd/bundles/hotfixer/*/patching-tool/patches
+	fi
 	read -p 'Please specify download link for the hotfix: ' hotfixdl
-	wget -c -N --user=calum.ragan --ask-password $hotfixdl -P $cwd/bundles/hotfixer/*/patching-tool/patches
+	wget -c -N --user={username} --ask-password $hotfixdl -P $cwd/bundles/hotfixer/*/patching-tool/patches
 	$cwd/bundles/hotfixer/*/patching-tool/patching-tool.sh install 
 
 	# Deleting Data and Logs folders and cleaning DB
 	cd $cwd/bundles/hotfixer/*/
 	echo -e "\e[44mDeleting the logs and data folder\e[0m"
 	rm -r data
-	rm -r logs
+	if [ -d logs ]; then rm -r logs; fi 
 	echo -e "\e[44mRunning the cleandb command-1\e[0m"
 	cleandb
 }
+
 #### Help documentation
 
 usage ()
@@ -383,8 +329,6 @@ usage ()
 	dl71               - Downloads the 7.1 CE GA3
 	dlmaster           - Downloads the latest master
 	dlhf			   - Asks tester if they require a fixpack
-	dlhfpt			   - Sets up a hotfix download that requires a fixpack
-	dlhfnpt			   - Sets up a hotfix download that doesn't require a fixpack
 	rstaging           - Sets up remote staging where remote is 8080 and live is 9080
